+++ test/dynamo/test_optimizations.py
@@ -8,6 +8,7 @@ import torch._dynamo.backends.ipex
 import torch._dynamo.test_case
 from torch._dynamo.backends.ipex import has_ipex
 from torch._dynamo.backends.onnxrt import has_onnxruntime
+from torch._dynamo.backends.tvm import has_tvm
 from torch._dynamo.testing import same
 
 
@@ -155,6 +156,10 @@ class TestOptimizations(torch._dynamo.test_case.TestCase):
     @unittest.skipIf(not has_onnxruntime(), "requires onnxruntime")
     def test_onnxrt(self):
         self._check_backend_works("onnxrt")
+
+    @unittest.skipIf(not has_tvm(), "requires tvm")
+    def test_tvm(self):
+        self._check_backend_works("tvm")
 
 
 class NormalizeIRTests(torch._dynamo.test_case.TestCase):
